+++ src/gcm.ts
@@ -10,23 +10,25 @@ import type * as Types from './types'
 const REGISTER_URL = 'https://android.clients.google.com/c2dm/register3'
 const CHECKIN_URL = 'https://android.clients.google.com/checkin'
 
+function parseLong(number: bigint, unsigned?: boolean | number, radix?: number): Long {
+    return Long.fromString(number.toString(), unsigned, radix)
+}
+
 export default async (config: Types.ClientConfig): Promise<Types.GcmData> => {
     const options = await checkIn(config)
     const credentials = await doRegister(options, config)
     return credentials
 }
 
-export async function checkIn(config: Types.ClientConfig): Promise<Types.GcmData> {
-    const body = await request<ArrayBuffer>({
-        ...config.axiosConfig,
-        url: CHECKIN_URL,
+export async function checkIn(config: Types.ClientConfig): Promise<Types.GcmCheckinResponse> {
+    const body = await (await request(CHECKIN_URL, {
         method: 'POST',
         headers: {
             'Content-Type': 'application/x-protobuf',
         },
-        data: prepareCheckinBuffer(config),
-        responseType: 'arraybuffer',
-    })
+        body: prepareCheckinBuffer(config),
+    })).arrayBuffer()
+    
 
     const AndroidCheckinResponse = Protos.checkin_proto.AndroidCheckinResponse
     const message = AndroidCheckinResponse.decode(new Uint8Array(body))
@@ -42,7 +44,7 @@ export async function checkIn(config: Types.ClientConfig): Promise<Types.GcmData
     }
 }
 
-async function doRegister({ androidId, securityToken }: Types.GcmData, config: Types.ClientConfig): Promise<Types.GcmData> {
+async function doRegister({ androidId, securityToken }: Types.GcmCheckinResponse, config: Types.ClientConfig): Promise<Types.GcmData> {
     const appId = `wp:${config.bundleId}#${randomUUID()}`
     const body = (new URLSearchParams({
         app: 'org.chromium.linux',
@@ -51,7 +53,7 @@ async function doRegister({ androidId, securityToken }: Types.GcmData, config: T
         sender: config.vapidKey,
     })).toString()
 
-    const response = await postRegister({ androidId, securityToken, body, axiosConfig: config.axiosConfig })
+    const response = await postRegister({ androidId, securityToken, body })
     const token = response.split('=')[1]
 
     return {
@@ -63,23 +65,20 @@ async function doRegister({ androidId, securityToken }: Types.GcmData, config: T
 }
 
 
-async function postRegister({ androidId, securityToken, body, retry = 0, axiosConfig }: {
+async function postRegister({ androidId, securityToken, body, retry = 0 }: {
     androidId: Types.GcmData['androidId']
     securityToken: Types.GcmData['securityToken']
     body: string
     retry?: number
-    axiosConfig: Types.ClientConfig['axiosConfig']
 }): Promise<string> {
-    const response = await request<string>({
-        ...axiosConfig,
-        url: REGISTER_URL,
+    const response = await (await request(REGISTER_URL, {
         method: 'POST',
         headers: {
             Authorization: `AidLogin ${androidId}:${securityToken}`,
             'Content-Type': 'application/x-www-form-urlencoded',
         },
-        data: body,
-    })
+        body,
+    })).text()
 
     if (response.includes('Error')) {
         Logger.warn(`Register request has failed with ${response}`)
@@ -89,7 +88,7 @@ async function postRegister({ androidId, securityToken, body, retry = 0, axiosCo
 
         Logger.warn(`Retry... ${retry + 1}`)
         await delay(1000)
-        return postRegister({ androidId, securityToken, body, retry: retry + 1, axiosConfig })
+        return postRegister({ androidId, securityToken, body, retry: retry + 1 })
     }
 
     return response
@@ -99,19 +98,36 @@ function prepareCheckinBuffer(config: Types.ClientConfig) {
     const gcm = config.credentials?.gcm
     const AndroidCheckinRequest = Protos.checkin_proto.AndroidCheckinRequest
 
-    const payload = {
-        userSerialNumber: 0,
+    const payload: Protos.checkin_proto.IAndroidCheckinRequest = {
+        accountCookie: [],
         checkin: {
-            type: 3,
+            cellOperator: '', // Optional
             chromeBuild: {
-                platform: 2,
-                chromeVersion: '63.0.3234.0',
-                channel: 1,
+                platform: config.chromePlatform || Protos.checkin_proto.ChromeBuildProto.Platform.PLATFORM_MAC,
+                chromeVersion: config.chromeVersion || '63.0.3234.0',
+                channel: config.chromeChannel || Protos.checkin_proto.ChromeBuildProto.Channel.CHANNEL_STABLE,
             },
+            type: Protos.checkin_proto.DeviceType.DEVICE_CHROME_BROWSER,
+            lastCheckinMsec: parseLong(0n), // TODO
+            roaming: '', // Optional
+            simOperator: '', // Optional
+            userNumber: 0, // Optional
         },
-        version: 3,
+        desiredBuild: '', // Optional
+        digest: '', // Optional
+        fragment: 0,
         id: gcm?.androidId ? Long.fromString(gcm.androidId) : undefined,
+        locale: '', // Optional
+        loggingId: parseLong(0n), // Optional
+        macAddr: [],
+        macAddrType: [],
+        marketCheckin: '', // Optional
+        otaCert: [],
         securityToken: gcm?.securityToken ? Long.fromString(gcm?.securityToken, true) : undefined,
+        timeZone: config.timeZone || 'Europe/Prague',
+        userName: '', // Optional
+        userSerialNumber: 0, // TODO
+        version: 3, // TODO - ???
     }
 
     const errMsg = AndroidCheckinRequest.verify(payload)