+++ src/index.ts
@@ -1,0 +1,9 @@
+import PushReceiver from "./client"
+import PushSender from "./sender"
+
+export {
+    PushReceiver,
+    PushSender,
+}
+
+export default PushReceiver