+++ src/utils/logger.ts
@@ -1,32 +1,17 @@
-import { LogLevels } from '../constants'
-
-let logLevel = LogLevels.NONE
-
-export const setLogLevel = (newLogLevel: keyof typeof LogLevels): void => {
-    logLevel = LogLevels[newLogLevel]
-}
-
-export const verbose = (...args: unknown[]): void => {
-    if (logLevel < LogLevels.VERBOSE) return
-    console.log('[PUSH_RECEIVER_VERBOSE]', ...args)
-}
-
-export const debug = (...args: unknown[]): void => {
-    if (logLevel < LogLevels.DEBUG) return
-    console.log('[PUSH_RECEIVER_DEBUG]', ...args)
-}
-
-export const warn = (...args: unknown[]): void => {
-    if (logLevel === LogLevels.NONE) return
-    console.warn('[PUSH_RECEIVER_WARNING]', ...args)
-}
-
-export const error = console.error
+let debugEnabled = false
 
 export default {
-    setLogLevel,
-    verbose,
-    debug,
-    warn,
-    error
+    setDebug: (enabled: boolean) => {
+        debugEnabled = Boolean(enabled)
+    },
+    log: (...args: unknown[]): void => {
+        console.log(...args)
+    },
+    debug: (...args: unknown[]): void => {
+        console.debug(...args)
+    },
+    warn: (...args: unknown[]): void => {
+        console.warn(...args)
+    },
+    error: console.error,
 }