+++ src/utils/request.ts
@@ -1,6 +1,6 @@
-import axios, { AxiosRequestConfig } from 'axios'
 import delay from './timeout'
 import Logger from './logger'
+import type { ClientConfig } from '../types'
 
 // In seconds
 const MAX_RETRY_TIMEOUT = 15
@@ -8,21 +8,26 @@ const MAX_RETRY_TIMEOUT = 15
 // Step in seconds
 const RETRY_STEP = 5
 
-export default function requestWithRety<T>(options: AxiosRequestConfig): Promise<T> {
-  return retry<T>(0, options)
+export default function requestWithRety(url: string, options?: globalThis.RequestInit, maxRetries = 3): Promise<Response> {
+    return retry(0, url, options, maxRetries)
 }
 
-async function retry<T>(retryCount = 0, options: AxiosRequestConfig): Promise<T> {
-  try {
-    const response = await axios(options)
-    return response.data
-  } catch (e) {
-    const timeout = Math.min(retryCount * RETRY_STEP, MAX_RETRY_TIMEOUT)
-    Logger.verbose(`Request failed : ${e.message}`)
-    Logger.verbose(`Retrying in ${timeout} seconds`)
+async function retry(retryCount = 0, url: string, options?: globalThis.RequestInit, maxRetries = 3): Promise<Response> {
+    try {
+        return await fetch(url, options)
+    } catch (e) {
+        const timeout = Math.min(retryCount * RETRY_STEP, MAX_RETRY_TIMEOUT)
+        Logger.debug(`Request failed : ${e.message}`)
+        Logger.debug(`Retrying in ${timeout} seconds`)
 
-    await delay(timeout * 1000)
+        if (retryCount >= maxRetries) throw e
 
-    return retry(retryCount + 1, options)
-  }
+        await delay(timeout * 1000)
+
+        return retry(retryCount + 1, url, options)
+    }
 }
+
+export const getEndpoint = (config: ClientConfig, baseUrl: string, path = '') => (
+    `${baseUrl.endsWith('/') ? baseUrl : `${baseUrl}/`}projects/${config.firebase.projectId}/${path}`
+)