+++ test/keys.template.js
@@ -2,4 +2,4 @@
 module.exports = {
   SENDER_ID  : 'Your sender id here',
   SERVER_KEY : 'Your server key here',
-};
+}